Add projects page with tag filtering
A projects page has been added, featuring a filter for the tags which
displays only projects with the given tag. Currently only a single
filter can be enabled at once, although in the future when there are
more projects this will be changed to toggling multiple, to more easily
refine the list of displayed projects.