Add navbar to home page
A navigation bar has been added to the homepage, placing links to the
projects, about and contact page up the top, the latter two of which do
not yet exist. These pages are already accessible via buttons on the
page, but this makes them more visible and easier to get to.